+++ docs/releases.rst
@@ -127,7 +127,15 @@ is now:
         f_call:
         -  "{F_result} = {F_C_call}({F_arg_c_call})"              
 
+* Likewise, some fields were renamed for Typemaps.
 
+=============   =============
+Old Name        New Name
+=============   =============
+f_c_module      i_module
+f_c_module_line i_module_line
+f_c_type        i_type
+           
 * Added format field *f_c_suffix*. Used in format fields
   *C_name_template* and *F_C_name_template* to allow Fortran wrapper
   *to call a C function with additional mangling such as